Move LogicFrame and LogicFrameSpan into OpenSage.SimCore.Ticking
The frame counter types are sim substrate (api-freeze-v1 F6), so they leave
OpenSage.Logic.Object for SimCore. The integer core is analyzer-clean; the
float conveniences the unmigrated float sim still calls live on in a
SIMCORE-EXEMPT FloatCompat file and are deleted with their callers per F11.
The Percentage/SageGame overloads could not follow (types not visible from
SimCore); their few call sites now spell the same semantics through the
float compat surface. Consumers keep compiling via global type aliases;
OpenSage.Game now references OpenSage.SimCore.

+++ src/OpenSage.Game/Logic/Object/Behaviors/GateOpenAndCloseBehavior.cs
@@ -70,7 +70,7 @@ public class GateOpenAndCloseBehavior : UpdateModule
                 GameObject.ModelConditionFlags.Set(ModelConditionFlag.Door1Closing, false);
                 GameObject.ModelConditionFlags.Set(ModelConditionFlag.Door1Opening, true);
                 _toggleFinishedTime = currentFrame + _moduleData.ResetTime;
-                _pathingToggleTime = currentFrame + (_moduleData.ResetTime * _moduleData.PercentOpenForPathing);
+                _pathingToggleTime = currentFrame + (_moduleData.ResetTime * (float)_moduleData.PercentOpenForPathing);
 
                 _openingSoundLoop = audioSystem.PlayAudioEvent(GameObject, _moduleData.SoundOpeningGateLoop?.Value, true);
                 _finishedSoundTime = currentFrame + _moduleData.TimeBeforePlayingOpenSound;
@@ -108,7 +108,7 @@ public class GateOpenAndCloseBehavior : UpdateModule
                 GameObject.ModelConditionFlags.Set(ModelConditionFlag.Door1Opening, false);
                 GameObject.ModelConditionFlags.Set(ModelConditionFlag.Door1Closing, true);
                 _toggleFinishedTime = currentFrame + _moduleData.ResetTime;
-                _pathingToggleTime = currentFrame + (_moduleData.ResetTime * _moduleData.PercentOpenForPathing);
+                _pathingToggleTime = currentFrame + (_moduleData.ResetTime * (float)_moduleData.PercentOpenForPathing);
 
                 _closingSoundLoop = audioSystem.PlayAudioEvent(GameObject, _moduleData.SoundClosingGateLoop?.Value, true);
                 _finishedSoundTime = currentFrame + _moduleData.TimeBeforePlayingClosedSound;
